+++ src/server/database/model/bottle.go
@@ -34,14 +34,14 @@ type Bottle struct {
 	Category  *string `json:"category"`
 	BrandID   uint64  `json:"brand_id" gorm:"not null"`
 	BottlerID *uint64 `json:"bottler_id"`
-	StatedAge *uint   `json:"stated_age"`
+	StatedAge *uint64 `json:"stated_age"`
 
 	Description   *string         `json:"description"`
 	TastingNotes  *datatypes.JSON `json:"tasting_notes"`
 	SuggestedTags *pq.StringArray `json:"suggested_tags" gorm:"type:string[];default:[];not null"`
 
 	AvgRating     *float64 `json:"avg_rating"`
-	TotalTastings uint     `json:"total_tastings" gorm:"default:0;not null"`
+	TotalTastings uint64   `json:"total_tastings" gorm:"default:0;not null"`
 
 	CreatedAt   time.Time `json:"created_at"`
 	CreatedByID uint64    `json:"created_by_id"`
